requirements.txt
absl-py==0.7.0
alabaster==0.7.12
albumentations==0.2.0
anaconda-client==1.2.2
anaconda-project==0.8.2
appnope==0.1.0
appscript==1.0.1
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
atomicwrites==1.2.1
attrs==19.1.0
audioread==2.1.6
autopep8==1.4.3
Babel==2.6.0
backcall==0.1.0
backports-abc==0.5
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
boto3==1.9.120
botocore==1.12.120
bz2file==0.98
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
clyent==1.2.1
cmake==3.13.3
colorama==0.3.9
conda==4.3.16
constants==0.6.0
contextlib2==0.5.5
Corpora==1.0
cryptography==2.6.1
CVXcanon==0.1.1
cycler==0.10.0
Cython==0.29.2
cytoolz==0.9.0.1
dask==1.1.4
datashape==0.5.2
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.8.2
Django==2.1.7
dlib==19.16.0
docker==3.7.1
docker-pycreds==0.4.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition==1.2.3
face-recognition-models==0.3.0
fastcache==1.0.2
Flask==1.0.2
Flask-Cors==3.0.7
gast==0.2.2
gensim==3.7.1
gevent==1.4.0
gmpy2==2.0.8
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-images-download==2.5.0
graphviz==0.10.1
greenlet==0.4.15
grpcio==1.16.1
grpcio-tools==1.19.0
h5py==2.8.0
HeapDict==1.0.0
html5lib==1.0.1
httplib2==0.12.1
idna==2.8
image==1.5.27
imageio==2.5.0
imageio-ffmpeg==0.2.0
imagesize==1.1.0
imgaug==0.2.8
importlib-metadata==0.7
intel-openmp==2019.0
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itsdangerous==1.1.0
Janome==0.3.7
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.4
jupyter-nbextensions-configurator==0.4.1
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==17.1.1
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lesscpy==0.13.0
librosa==0.6.3
lightgbm==2.2.3
llvmlite==0.27.0
locket==0.2.0
lxml==4.3.2
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
missingno==0.4.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mmdnn==0.2.4
more-itertools==5.0.0
moviepy==1.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.42.1
numpy==1.15.4
numpydoc==0.8.0
oauth2client==4.1.3
olefile==0.46
opencv-python==4.0.0.21
openpyxl==2.5.12
packaging==18.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pipenv==2018.11.26
plotly==3.7.0
pluggy==0.8.1
ply==3.11
proglog==0.1.9
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.6.1
psutil==5.4.8
ptyprocess==0.6.0
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
PyAutoGUI==0.9.41
pybind11==2.2.4
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
PyDrive==1.3.1
pydub==0.23.1
pyflakes==2.0.0
pygame==1.9.5
PyGetWindow==0.0.4
Pygments==2.3.1
pylint==2.2.2
PyMsgBox==1.0.6
pyobjc-core==5.1.2
pyobjc-framework-Cocoa==5.1.2
pyobjc-framework-Quartz==5.1.2
pyodbc==4.0.25
pyOpenSSL==19.0.0
pyparsing==2.3.1
PyQt5==5.9.2
PyRect==0.1.4
pyrsistent==0.14.11
PyScreeze==0.1.19
PySocks==1.6.8
pytest==4.1.1
python-dateutil==2.8.0
pytube==9.4.0
PyTweening==1.0.3
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.5
qtconsole==4.4.3
QtPy==1.5.2
requests==2.21.0
resampy==0.2.1
retrying==1.3.3
rope==0.11.0
rsa==4.0
ruamel.yaml==0.15.87
s3transfer==0.2.0
scandir==1.5
scikit-image==0.14.2
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
simplegeneric==0.8.1
singledispatch==3.4.0.3
sip==4.19.8
six==1.12.0
sklearn==0.0
smart-open==1.8.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.7.1
Sphinx==1.8.2
sphinxcontrib-websupport==1.1.0
spyder==3.3.2
spyder-kernels==0.3.0
SQLAlchemy==1.2.16
statsmodels==0.9.0
stopwords==0.1.3
sympy==1.3
tblib==1.3.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
typed-ast==1.1.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.1
utils==0.9.0
virtualenv==16.4.3
virtualenv-clone==0.5.1
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wordcloud==1.5.0
wordnet==0.0.1b2
wrapt==1.11.0
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.2
xlwings==0.15.1
xlwt==1.3.0
youtube-dl==2019.4.7
zict==0.1.3
